+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change Log
 
-All notable changes to the "contentsgenerator" extension will be documented in this file.
+All notable changes to the "TOC generator for Markdown" extension will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
+
+## [Release 1.0.2] - 2025-1-3
+### Changed
+- command name
+
+### Added
+- tocgenerator.generateContentsTablegh: "create ToC for GitHub MD"
 
 ## [Release 1.0.1] - 2025-1-1
 ### Changed